Feat: 추가정보 입력 레이아웃 정리
