+++ packages/web/src/app/@modal/(.)additional-info/page.tsx
@@ -1,10 +1,13 @@
 import Modal from '@/components/modal/modal'
-import AdditionalInfoLayout from '@/containers/Additional/AdditionalLayout'
+import AdditionalInfoLayout from '@/containers/additional-info/AdditionalInfoLayout'
 
 export default function Page() {
   return (
     <Modal>
-      <AdditionalInfoLayout className="" />
+      <AdditionalInfoLayout
+        className={`h-full py-[50px]
+                  flex flex-col items-center justify-around`}
+      />
     </Modal>
   )
 }
