+++ packages/web/src/app/page.tsx
@@ -1,3 +1,5 @@
+import Link from 'next/link'
+
 import { ArrowIcon, Button, Input } from '@packages/ui'
 
 import Navbar from '@/components/Navbar'
@@ -13,6 +15,7 @@ export default function Page() {
       <main>
         <h1>컴포넌트dd</h1>
         <p>테스트</p>
+        <Link href="additional-info">모달얍</Link>
       </main>
     </>
   )
