+++ packages/web/src/components/form/formLayout.tsx
@@ -1,9 +1,33 @@
-interface Props {
-  children: React.ReactNode
+import styles from './form.module.css'
+
+const Legend = ({
+  title,
+  description,
+}: {
+  title?: string
+  description?: string
+}) => {
+  return (
+    <div className={`${styles['title-container']}`}>
+      <h3>{title}</h3>
+      <p>{description}</p>
+    </div>
+  )
 }
 
-const FormItems = ({ children }: Props) => {
-  return { children }
+export default function FormLayout({
+  className,
+  children,
+}: {
+  className?: string
+  children?: React.ReactNode
+}) {
+  return (
+    <section className={className}>
+      <Legend />
+      {children}
+    </section>
+  )
 }
 
-export default FormItems
+FormLayout.Legend = Legend
