+++ packages/web/src/components/modal/modal.tsx
@@ -8,7 +8,13 @@ import { CloseIcon } from '@packages/ui'
 
 import styles from './modal.module.css'
 
-function Modal({ children }: { children: React.ReactNode }) {
+function Modal({
+  className,
+  children,
+}: {
+  className?: string
+  children: React.ReactNode
+}) {
   const router = useRouter()
   const dialogRef = useRef<ElementRef<'dialog'>>(null)
 
@@ -26,6 +32,7 @@ function Modal({ children }: { children: React.ReactNode }) {
     <div className="absolute bg-[rgba(0,0,0,0.7)] flex justify-center items-center z-[1000] inset-0">
       <dialog
         ref={dialogRef}
+        // TODO: className으로 모달 크기 프롭받기
         className={`w-[90%] h-[90%] sm:w-[80%] sm:h-[80%] 
                   md:w-[60%] md:h-[60%] lg:w-[50%] lg:h-[60%]
                   flex flex-col items-center justify-around
@@ -33,6 +40,8 @@ function Modal({ children }: { children: React.ReactNode }) {
                   bg-[url('/images/BG.svg')] bg-cover bg-center bg-no-repeat 
                   overflow-hidden
                   rounded-[10px]
+                  
+                  ${className}
                   `}
         onClose={onDismiss}
       >
